[gemspec] Work around dependabot bug by using 'File.public_send' (#38)
... rather than `File.read` to read the Ruby version from the
`.ruby-version` file.

The issue is that Dependabot overwrites all `File.read` content in the
gemspec with the gem's own version number (I guess on the assumption
that `File.read` might be used to read the value of `spec.version`, i.e.
the gem's own version number, from a version file). I think that
basically happens here:
https://github.com/dependabot/dependabot-core/blob/42d95a29/bundler/lib/dependabot/bundler/file_updater/gemspec_sanitizer.rb/#L369-L371